refactor processing activity unit level metadata

+++ gdpr/src/main/java/com/kairos/controller/data_inventory/processing_activity/OrganizationAccessorPartyController.java
@@ -45,7 +45,7 @@ public class OrganizationAccessorPartyController {
 
     @ApiOperation("get AccessorParty by id")
     @GetMapping("/accessor_party/{accessorPartyId}")
-    public ResponseEntity<Object> getAccessorPartyById(@PathVariable Long unitId, @PathVariable BigInteger accessorPartyId) {
+    public ResponseEntity<Object> getAccessorPartyById(@PathVariable Long unitId, @PathVariable Long accessorPartyId) {
         if (unitId == null) {
             return ResponseHandler.invalidResponse(HttpStatus.BAD_REQUEST, false, "organization id can't be null");
         }
@@ -76,7 +76,7 @@ public class OrganizationAccessorPartyController {
 
     @ApiOperation("update AccessorParty by id")
     @PutMapping("/accessor_party/{accessorPartyId}")
-    public ResponseEntity<Object> updateAccessorParty(@PathVariable Long unitId, @PathVariable BigInteger accessorPartyId, @Valid @RequestBody AccessorPartyDTO accessorParty) {
+    public ResponseEntity<Object> updateAccessorParty(@PathVariable Long unitId, @PathVariable Long accessorPartyId, @Valid @RequestBody AccessorPartyDTO accessorParty) {
         if (unitId == null) {
             return ResponseHandler.invalidResponse(HttpStatus.BAD_REQUEST, false, "organization id can't be null");
