+++ gdpr/src/main/java/com/kairos/controller/data_inventory/processing_activity/OrganizationProcessingLegalBasisController.java
@@ -44,7 +44,7 @@ public class OrganizationProcessingLegalBasisController {
 
     @ApiOperation("get ProcessingLegalBasis by id")
     @GetMapping("/legal_basis/{legalBasisId}")
-    public ResponseEntity<Object> getProcessingLegalBasis(@PathVariable Long unitId, @PathVariable BigInteger legalBasisId) {
+    public ResponseEntity<Object> getProcessingLegalBasis(@PathVariable Long unitId, @PathVariable Long legalBasisId) {
         if (unitId == null) {
             return ResponseHandler.invalidResponse(HttpStatus.BAD_REQUEST, false, "organization id can't be null");
         }
@@ -74,7 +74,7 @@ public class OrganizationProcessingLegalBasisController {
 
     @ApiOperation("update ProcessingLegalBasis by id")
     @PutMapping("/legal_basis/{legalBasisId}")
-    public ResponseEntity<Object> updateProcessingLegalBasis(@PathVariable Long unitId, @PathVariable BigInteger legalBasisId, @Valid @RequestBody ProcessingLegalBasisDTO legalBasis) {
+    public ResponseEntity<Object> updateProcessingLegalBasis(@PathVariable Long unitId, @PathVariable Long legalBasisId, @Valid @RequestBody ProcessingLegalBasisDTO legalBasis) {
         if (unitId == null) {
             return ResponseHandler.invalidResponse(HttpStatus.BAD_REQUEST, false, "organization id can't be null");
         }
